+++ packages/react-scripts/config/jest/babelTransform.js
@@ -1,3 +1,4 @@
+// @remove-file-on-eject
 /**
  * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
  *
