+++ packages/react-scripts/scripts/utils/createJestConfig.js
@@ -1,3 +1,4 @@
+// @remove-file-on-eject
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
  * All rights reserved.
@@ -7,10 +8,8 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
-// Note: this file does not exist after ejecting.
-
 const fs = require('fs');
-const paths = require('../config/paths');
+const paths = require('../../config/paths');
 
 module.exports = (resolve, rootDir, isEjecting) => {
   // Use this instead of `paths.testsSetup` to avoid putting
